Signing up and singing in implemented
