+++ src/main/java/com/pbl/yourside/security/reponse/ResponseMessage.java
@@ -1,4 +1,4 @@
-package com.pbl.yourside.security.sevices.reponse;
+package com.pbl.yourside.security.reponse;
 
 public class ResponseMessage {
     private String message;
