+++ src/main/java/com/pbl/yourside/security/request/SignUpForm.java
@@ -1,13 +1,15 @@
-package com.pbl.yourside.security.sevices.request;
+package com.pbl.yourside.security.request;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
+import java.util.Set;
 
-public class LoginForm {
-
+public class SignUpForm {
     @NotBlank
-    @Size(min=3, max = 60)
+    @Size(min = 3, max = 50)
     private String username;
+
+    private Set<String> role;
 
     @NotBlank
     @Size(min = 6, max = 40)
@@ -19,6 +21,14 @@ public class LoginForm {
 
     public void setUsername(String username) {
         this.username = username;
+    }
+
+    public Set<String> getRole() {
+        return role;
+    }
+
+    public void setRole(Set<String> role) {
+        this.role = role;
     }
 
     public String getPassword() {
